# Pytracer improvements