@nedbat 2022-05-30
I'm surprised it's worthwhile to use an attribute to avoid a dict lookup.  I would have thought the extra bytecodes would be worse than just doing the dict lookup, but I see the results.  This seems to be about a 10% improvement on my benchmark.