@nedbat 2022-06-02
This is now released as part of [coverage 6.4.1](https://pypi.org/project/coverage/6.4.1).